Add :print directive for customizable repl printing
Refs #3177. The command

    > :print MyModule.print

will replace the repl's printing function (by default:
PSCI.Support.eval) with MyModule.print, if the latter exists and appears
to work. The referenced function should have a type of the form

    forall a. (C a) => a -> Effect _

i.e., it should accept a wide variety of input types, and return
an `Effect T` for some type `T`, whose effects involve printing the
argument to the console. The `T` value is discarded.

It is not yet possible to configure the repl's printing function from
the .purs-repl file.